+++ src/tree/hex/edge_template_1/mod.rs
@@ -1,7 +1,7 @@
 use super::super::{
-    Coordinates, HexConnectivity, NODE_NUMBERING_OFFSET, NUM_OCTANTS, NodeMap, Octree,
+    Cell, Coordinates, HexConnectivity, NODE_NUMBERING_OFFSET, NodeMap, Octree, mirror_face,
 };
-use conspire::math::{TensorVec, tensor_rank_1};
+use conspire::math::{TensorRank1, TensorVec, tensor_rank_1};
 
 pub fn apply(
     cells_nodes: &[usize],
@@ -11,367 +11,316 @@ pub fn apply(
     element_node_connectivity: &mut HexConnectivity,
     nodal_coordinates: &mut Coordinates,
 ) {
-    tree.iter()
-        .filter_map(|cell| tree.cell_contains_leaves(cell))
-        .for_each(|(cell_subcells, _)| {
-            template(
-                0,
-                1,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            );
-            template(
-                0,
-                2,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            );
-            template(
-                0,
-                4,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            );
-            template(
-                1,
-                3,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            );
-            template(
-                1,
-                5,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            );
-            template(
-                2,
-                3,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            );
-            template(
-                2,
-                6,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            );
-            template(
-                3,
-                7,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            );
-            template(
-                4,
-                5,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            );
-            template(
-                4,
-                6,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            );
-            template(
-                5,
-                7,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            );
-            template(
-                6,
-                7,
-                cells_nodes,
-                cell_subcells,
-                nodes_map,
-                node_index,
-                tree,
-                element_node_connectivity,
-                nodal_coordinates,
-            )
-        })
+    tree.iter().for_each(|cell| {
+        template(
+            0,
+            1,
+            cell,
+            cells_nodes,
+            nodes_map,
+            node_index,
+            tree,
+            element_node_connectivity,
+            nodal_coordinates,
+        );
+        template(
+            0,
+            4,
+            cell,
+            cells_nodes,
+            nodes_map,
+            node_index,
+            tree,
+            element_node_connectivity,
+            nodal_coordinates,
+        );
+        template(
+            1,
+            2,
+            cell,
+            cells_nodes,
+            nodes_map,
+            node_index,
+            tree,
+            element_node_connectivity,
+            nodal_coordinates,
+        );
+        template(
+            1,
+            4,
+            cell,
+            cells_nodes,
+            nodes_map,
+            node_index,
+            tree,
+            element_node_connectivity,
+            nodal_coordinates,
+        );
+        template(
+            2,
+            3,
+            cell,
+            cells_nodes,
+            nodes_map,
+            node_index,
+            tree,
+            element_node_connectivity,
+            nodal_coordinates,
+        );
+        template(
+            2,
+            5,
+            cell,
+            cells_nodes,
+            nodes_map,
+            node_index,
+            tree,
+            element_node_connectivity,
+            nodal_coordinates,
+        );
+        template(
+            3,
+            0,
+            cell,
+            cells_nodes,
+            nodes_map,
+            node_index,
+            tree,
+            element_node_connectivity,
+            nodal_coordinates,
+        );
+        template(
+            4,
+            2,
+            cell,
+            cells_nodes,
+            nodes_map,
+            node_index,
+            tree,
+            element_node_connectivity,
+            nodal_coordinates,
+        );
+        template(
+            4,
+            3,
+            cell,
+            cells_nodes,
+            nodes_map,
+            node_index,
+            tree,
+            element_node_connectivity,
+            nodal_coordinates,
+        );
+        template(
+            5,
+            0,
+            cell,
+            cells_nodes,
+            nodes_map,
+            node_index,
+            tree,
+            element_node_connectivity,
+            nodal_coordinates,
+        );
+        template(
+            5,
+            1,
+            cell,
+            cells_nodes,
+            nodes_map,
+            node_index,
+            tree,
+            element_node_connectivity,
+            nodal_coordinates,
+        );
+        template(
+            3,
+            5,
+            cell,
+            cells_nodes,
+            nodes_map,
+            node_index,
+            tree,
+            element_node_connectivity,
+            nodal_coordinates,
+        );
+    })
 }
 
 #[allow(clippy::too_many_arguments)]
 fn template(
-    subcell_a: usize,
-    subcell_b: usize,
+    face_m: usize,
+    face_n: usize,
+    cell: &Cell,
     cells_nodes: &[usize],
-    cell_subcells: &[usize; NUM_OCTANTS],
     nodes_map: &mut NodeMap,
     node_index: &mut usize,
     tree: &Octree,
     element_node_connectivity: &mut HexConnectivity,
     nodal_coordinates: &mut Coordinates,
 ) {
-    let (face_m, face_n, subcell_c, subcell_d, subcell_e, subcell_f, flip, direction) =
-        match (subcell_a, subcell_b) {
-            (0, 1) => (0, 4, 2, 4, 3, 5, false, tensor_rank_1([0.0, 1.0, 0.0])),
-            (0, 2) => (3, 4, 1, 4, 3, 6, true, tensor_rank_1([1.0, 0.0, 0.0])),
-            (0, 4) => (3, 0, 1, 2, 5, 6, false, tensor_rank_1([1.0, 0.0, 0.0])),
-            (1, 3) => (1, 4, 0, 5, 2, 7, false, tensor_rank_1([-1.0, 0.0, 0.0])),
-            (1, 5) => (0, 1, 3, 0, 7, 4, false, tensor_rank_1([0.0, 1.0, 0.0])),
-            (2, 3) => (2, 4, 0, 6, 1, 7, true, tensor_rank_1([0.0, -1.0, 0.0])),
-            (2, 6) => (2, 3, 0, 3, 4, 7, false, tensor_rank_1([0.0, -1.0, 0.0])),
-            (3, 7) => (1, 2, 2, 1, 6, 5, false, tensor_rank_1([-1.0, 0.0, 0.0])),
-            (4, 5) => (5, 0, 0, 6, 1, 7, false, tensor_rank_1([0.0, 0.0, -1.0])),
-            (4, 6) => (5, 3, 0, 5, 2, 7, true, tensor_rank_1([0.0, 0.0, -1.0])),
-            (5, 7) => (5, 1, 1, 4, 3, 6, false, tensor_rank_1([0.0, 0.0, -1.0])),
-            (6, 7) => (5, 2, 2, 4, 3, 5, true, tensor_rank_1([0.0, 0.0, -1.0])),
+    let (
+        subcell_m_a,
+        subcell_m_b,
+        subcell_m_c,
+        subcell_m_d,
+        subcell_face_m_a,
+        subcell_face_m_b,
+        subcell_face_n_a,
+        subcell_face_n_b,
+        subcell_diag_mn_a,
+        subcell_diag_mn_b,
+    ) = match (face_m, face_n) {
+        (0, 1) => (7, 13, 5, 15, 3, 7, 0, 4, 2, 6),
+        (0, 4) => (1, 4, 0, 5, 2, 3, 4, 5, 6, 7),
+        (1, 2) => (7, 13, 5, 15, 2, 6, 1, 5, 0, 4),
+        (1, 4) => (1, 4, 0, 5, 0, 2, 5, 7, 4, 6),
+        (2, 3) => (2, 8, 0, 10, 0, 4, 3, 7, 1, 5),
+        (2, 5) => (11, 14, 10, 15, 4, 5, 2, 3, 0, 1),
+        (3, 0) => (2, 8, 0, 10, 1, 5, 2, 6, 3, 7),
+        (3, 5) => (11, 14, 10, 15, 5, 7, 0, 2, 1, 3),
+        (4, 2) => (11, 14, 10, 15, 6, 7, 0, 1, 4, 5),
+        (4, 3) => (2, 8, 0, 10, 4, 6, 1, 3, 5, 7),
+        (5, 0) => (1, 4, 0, 5, 0, 1, 6, 7, 2, 3),
+        (5, 1) => (7, 13, 5, 15, 1, 3, 4, 6, 0, 2),
+        _ => panic!(),
+    };
+    let directions: [TensorRank1<3, 1>; 2] = [face_m, face_n]
+        .iter()
+        .map(|face| match face {
+            0 => tensor_rank_1([0.0, -1.0, 0.0]),
+            1 => tensor_rank_1([1.0, 0.0, 0.0]),
+            2 => tensor_rank_1([0.0, 1.0, 0.0]),
+            3 => tensor_rank_1([-1.0, 0.0, 0.0]),
+            4 => tensor_rank_1([0.0, 0.0, -1.0]),
+            5 => tensor_rank_1([0.0, 0.0, 1.0]),
             _ => panic!(),
-        };
-    let subcell_a_faces = tree[cell_subcells[subcell_a]].get_faces();
-    if let Some(subcell_a_face_m) = subcell_a_faces[face_m] {
-        if let Some(subcell_a_face_n) = subcell_a_faces[face_n] {
-            if let Some((subcell_a_face_m_subcells, _)) =
-                tree.cell_contains_leaves(&tree[subcell_a_face_m])
-            {
-                if let Some((subcell_a_face_n_subcells, _)) =
-                    tree.cell_contains_leaves(&tree[subcell_a_face_n])
-                {
-                    if let Some(diagonal_a) =
-                        tree[subcell_a_face_m_subcells[subcell_c]].get_faces()[face_n]
+        })
+        .collect::<Vec<TensorRank1<3, 1>>>()
+        .try_into()
+        .unwrap();
+    if let Some(cell_face_m) = cell.get_faces()[face_m] {
+        if let Some(cell_face_n) = cell.get_faces()[face_n] {
+            if let Some(cell_diag_mn) = tree[cell_face_m].get_faces()[face_n] {
+                if let Some((subcells_face_m, _)) = tree.cell_contains_leaves(&tree[cell_face_m]) {
+                    if let Some((subcells_face_n, _)) =
+                        tree.cell_contains_leaves(&tree[cell_face_n])
                     {
-                        if let Some(subdiagonal_a) =
-                            tree[subcell_a_face_m_subcells[subcell_e]].get_faces()[face_n]
+                        if let Some((subcells_diag_mn, _)) =
+                            tree.cell_contains_leaves(&tree[cell_diag_mn])
                         {
-                            let subcell_b_faces = tree[cell_subcells[subcell_b]].get_faces();
-                            if let Some(subcell_b_face_m) = subcell_b_faces[face_m] {
-                                if let Some(subcell_b_face_n) = subcell_b_faces[face_n] {
-                                    if let Some((subcell_b_face_m_subcells, _)) =
-                                        tree.cell_contains_leaves(&tree[subcell_b_face_m])
-                                    {
-                                        if let Some((subcell_b_face_n_subcells, _)) =
-                                            tree.cell_contains_leaves(&tree[subcell_b_face_n])
-                                        {
-                                            if let Some(diagonal_b) = tree
-                                                [subcell_b_face_m_subcells[subcell_e]]
-                                                .get_faces()[face_n]
-                                            {
-                                                if let Some(subdiagonal_b) = tree
-                                                    [subcell_b_face_m_subcells[subcell_c]]
-                                                    .get_faces()[face_n]
-                                                {
-                                                    if cells_nodes[diagonal_a] == 0 {
-                                                        println!(
-                                                            "Cell diagonal_a={} is node 0, leaf is {}",
-                                                            diagonal_a,
-                                                            tree[diagonal_a].is_leaf()
-                                                        );
-                                                    } else {
-                                                        assert!(tree[diagonal_a].is_leaf())
-                                                    }
-                                                    if cells_nodes[diagonal_b] == 0 {
-                                                        println!(
-                                                            "Cell diagonal_b={} is node 0, leaf is {}",
-                                                            diagonal_b,
-                                                            tree[diagonal_b].is_leaf()
-                                                        );
-                                                    } else {
-                                                        assert!(tree[diagonal_b].is_leaf())
-                                                    }
-                                                    if cells_nodes[subdiagonal_a] == 0 {
-                                                        println!(
-                                                            "Cell subdiagonal_a={} is node 0, leaf is {}",
-                                                            subdiagonal_a,
-                                                            tree[subdiagonal_a].is_leaf()
-                                                        );
-                                                    } else {
-                                                        assert!(tree[subdiagonal_a].is_leaf())
-                                                    }
-                                                    if cells_nodes[subdiagonal_b] == 0 {
-                                                        println!(
-                                                            "Cell subdiagonal_b={} is node 0, leaf is {}",
-                                                            subdiagonal_b,
-                                                            tree[subdiagonal_b].is_leaf()
-                                                        );
-                                                    } else {
-                                                        assert!(tree[subdiagonal_b].is_leaf())
-                                                    }
-                                                    let lngth = *tree
-                                                        [subcell_a_face_m_subcells[subcell_e]]
-                                                        .get_lngth()
-                                                        as f64;
-                                                    nodal_coordinates.push(
-                                                        &nodal_coordinates[cells_nodes
-                                                            [subcell_a_face_m_subcells[subcell_e]]
-                                                            - NODE_NUMBERING_OFFSET]
-                                                            + &direction * lngth,
-                                                    );
-                                                    nodal_coordinates.push(
-                                                        &nodal_coordinates[cells_nodes
-                                                            [subcell_b_face_m_subcells[subcell_c]]
-                                                            - NODE_NUMBERING_OFFSET]
-                                                            + direction * lngth,
-                                                    );
-                                                    (0..2).for_each(|k| {
-                                                        assert!(nodes_map.insert(
-                                                            (
-                                                                (2.0 * nodal_coordinates[*node_index
-                                                                    + k
-                                                                    - NODE_NUMBERING_OFFSET][0])
-                                                                    as usize,
-                                                                (2.0 * nodal_coordinates[*node_index
-                                                                    + k
-                                                                    - NODE_NUMBERING_OFFSET][1])
-                                                                    as usize,
-                                                                (2.0 * nodal_coordinates[*node_index
-                                                                    + k
-                                                                    - NODE_NUMBERING_OFFSET][2])
-                                                                    as usize,
-                                                            ),
-                                                            *node_index + k,
-                                                        ).is_none(), "duplicate entry")
-                                                    });
-                                                    if flip {
-                                                        element_node_connectivity.push([
-                                                            *node_index,
-                                                            cells_nodes[subcell_a_face_m_subcells
-                                                                [subcell_e]],
-                                                            cells_nodes[subdiagonal_a],
-                                                            cells_nodes[subcell_a_face_n_subcells
-                                                                [subcell_f]],
-                                                            cells_nodes[cell_subcells[subcell_a]],
-                                                            cells_nodes[subcell_a_face_m_subcells
-                                                                [subcell_c]],
-                                                            cells_nodes[diagonal_a],
-                                                            cells_nodes[subcell_a_face_n_subcells
-                                                                [subcell_d]],
-                                                        ]);
-                                                        element_node_connectivity.push([
-                                                            *node_index + 1,
-                                                            cells_nodes[subcell_b_face_m_subcells
-                                                                [subcell_c]],
-                                                            cells_nodes[subdiagonal_b],
-                                                            cells_nodes[subcell_b_face_n_subcells
-                                                                [subcell_d]],
-                                                            *node_index,
-                                                            cells_nodes[subcell_a_face_m_subcells
-                                                                [subcell_e]],
-                                                            cells_nodes[subdiagonal_a],
-                                                            cells_nodes[subcell_a_face_n_subcells
-                                                                [subcell_f]],
-                                                        ]);
-                                                        element_node_connectivity.push([
-                                                            cells_nodes[cell_subcells[subcell_b]],
-                                                            cells_nodes[subcell_b_face_m_subcells
-                                                                [subcell_e]],
-                                                            cells_nodes[diagonal_b],
-                                                            cells_nodes[subcell_b_face_n_subcells
-                                                                [subcell_f]],
-                                                            *node_index + 1,
-                                                            cells_nodes[subcell_b_face_m_subcells
-                                                                [subcell_c]],
-                                                            cells_nodes[subdiagonal_b],
-                                                            cells_nodes[subcell_b_face_n_subcells
-                                                                [subcell_d]],
-                                                        ]);
-                                                    } else {
-                                                        element_node_connectivity.push([
-                                                            cells_nodes[cell_subcells[subcell_a]],
-                                                            cells_nodes[subcell_a_face_m_subcells
-                                                                [subcell_c]],
-                                                            cells_nodes[diagonal_a],
-                                                            cells_nodes[subcell_a_face_n_subcells
-                                                                [subcell_d]],
-                                                            *node_index,
-                                                            cells_nodes[subcell_a_face_m_subcells
-                                                                [subcell_e]],
-                                                            cells_nodes[subdiagonal_a],
-                                                            cells_nodes[subcell_a_face_n_subcells
-                                                                [subcell_f]],
-                                                        ]);
-                                                        element_node_connectivity.push([
-                                                            *node_index,
-                                                            cells_nodes[subcell_a_face_m_subcells
-                                                                [subcell_e]],
-                                                            cells_nodes[subdiagonal_a],
-                                                            cells_nodes[subcell_a_face_n_subcells
-                                                                [subcell_f]],
-                                                            *node_index + 1,
-                                                            cells_nodes[subcell_b_face_m_subcells
-                                                                [subcell_c]],
-                                                            cells_nodes[subdiagonal_b],
-                                                            cells_nodes[subcell_b_face_n_subcells
-                                                                [subcell_d]],
-                                                        ]);
-                                                        element_node_connectivity.push([
-                                                            *node_index + 1,
-                                                            cells_nodes[subcell_b_face_m_subcells
-                                                                [subcell_c]],
-                                                            cells_nodes[subdiagonal_b],
-                                                            cells_nodes[subcell_b_face_n_subcells
-                                                                [subcell_d]],
-                                                            cells_nodes[cell_subcells[subcell_b]],
-                                                            cells_nodes[subcell_b_face_m_subcells
-                                                                [subcell_e]],
-                                                            cells_nodes[diagonal_b],
-                                                            cells_nodes[subcell_b_face_n_subcells
-                                                                [subcell_f]],
-                                                        ]);
-                                                    }
-                                                    *node_index += 2;
-                                                }
-                                            }
-                                        }
-                                    }
+                            if let Some(subcells_m) =
+                                tree.cell_subcells_contain_leaves(cell, mirror_face(face_m))
+                            {
+                                if tree
+                                    .cell_subcells_contain_leaves(cell, mirror_face(face_n))
+                                    .is_some()
+                                {
+                                    let lngth = *tree[subcells_m[subcell_m_a]].get_lngth() as f64;
+                                    nodal_coordinates.push(
+                                        &nodal_coordinates[cells_nodes[subcells_m[subcell_m_a]]
+                                            - NODE_NUMBERING_OFFSET]
+                                            + &directions[0] * lngth,
+                                    );
+                                    nodal_coordinates.push(
+                                        &nodal_coordinates[cells_nodes[subcells_m[subcell_m_a]]
+                                            - NODE_NUMBERING_OFFSET]
+                                            + &directions[0] * lngth
+                                            + &directions[1] * lngth,
+                                    );
+                                    nodal_coordinates.push(
+                                        &nodal_coordinates[cells_nodes[subcells_m[subcell_m_a]]
+                                            - NODE_NUMBERING_OFFSET]
+                                            + &directions[1] * lngth,
+                                    );
+                                    nodal_coordinates.push(
+                                        &nodal_coordinates[cells_nodes[subcells_m[subcell_m_b]]
+                                            - NODE_NUMBERING_OFFSET]
+                                            + &directions[0] * lngth,
+                                    );
+                                    nodal_coordinates.push(
+                                        &nodal_coordinates[cells_nodes[subcells_m[subcell_m_b]]
+                                            - NODE_NUMBERING_OFFSET]
+                                            + &directions[0] * lngth
+                                            + &directions[1] * lngth,
+                                    );
+                                    nodal_coordinates.push(
+                                        &nodal_coordinates[cells_nodes[subcells_m[subcell_m_b]]
+                                            - NODE_NUMBERING_OFFSET]
+                                            + &directions[1] * lngth,
+                                    );
+                                    (0..6).for_each(|k| {
+                                        assert!(
+                                            nodes_map
+                                                .insert(
+                                                    (
+                                                        (2.0 * nodal_coordinates[*node_index + k
+                                                            - NODE_NUMBERING_OFFSET][0])
+                                                            as usize,
+                                                        (2.0 * nodal_coordinates[*node_index + k
+                                                            - NODE_NUMBERING_OFFSET][1])
+                                                            as usize,
+                                                        (2.0 * nodal_coordinates[*node_index + k
+                                                            - NODE_NUMBERING_OFFSET][2])
+                                                            as usize,
+                                                    ),
+                                                    *node_index + k,
+                                                )
+                                                .is_none(),
+                                            "duplicate entry"
+                                        )
+                                    });
+                                    element_node_connectivity.push([
+                                        cells_nodes[subcells_m[subcell_m_a]],
+                                        *node_index,
+                                        *node_index + 1,
+                                        *node_index + 2,
+                                        cells_nodes[subcells_m[subcell_m_b]],
+                                        *node_index + 3,
+                                        *node_index + 4,
+                                        *node_index + 5,
+                                    ]);
+                                    element_node_connectivity.push([
+                                        *node_index,
+                                        cells_nodes[subcells_face_m[subcell_face_m_a]],
+                                        cells_nodes[subcells_diag_mn[subcell_diag_mn_a]],
+                                        *node_index + 1,
+                                        *node_index + 3,
+                                        cells_nodes[subcells_face_m[subcell_face_m_b]],
+                                        cells_nodes[subcells_diag_mn[subcell_diag_mn_b]],
+                                        *node_index + 4,
+                                    ]);
+                                    element_node_connectivity.push([
+                                        *node_index + 1,
+                                        cells_nodes[subcells_diag_mn[subcell_diag_mn_a]],
+                                        cells_nodes[subcells_face_n[subcell_face_n_a]],
+                                        *node_index + 2,
+                                        *node_index + 4,
+                                        cells_nodes[subcells_diag_mn[subcell_diag_mn_b]],
+                                        cells_nodes[subcells_face_n[subcell_face_n_b]],
+                                        *node_index + 5,
+                                    ]);
+                                    element_node_connectivity.push([
+                                        cells_nodes[subcells_m[subcell_m_a]],
+                                        *node_index + 2,
+                                        *node_index + 1,
+                                        *node_index,
+                                        cells_nodes[subcells_m[subcell_m_c]],
+                                        cells_nodes[subcells_face_n[subcell_face_n_a]],
+                                        cells_nodes[subcells_diag_mn[subcell_diag_mn_a]],
+                                        cells_nodes[subcells_face_m[subcell_face_m_a]],
+                                    ]);
+                                    element_node_connectivity.push([
+                                        cells_nodes[subcells_m[subcell_m_b]],
+                                        *node_index + 3,
+                                        *node_index + 4,
+                                        *node_index + 5,
+                                        cells_nodes[subcells_m[subcell_m_d]],
+                                        cells_nodes[subcells_face_m[subcell_face_m_b]],
+                                        cells_nodes[subcells_diag_mn[subcell_diag_mn_b]],
+                                        cells_nodes[subcells_face_n[subcell_face_n_b]],
+                                    ]);
+                                    *node_index += 6;
                                 }
                             }
                         }
