+++ src/tree/hex/edge_template_2/mod.rs
@@ -1,332 +1,200 @@
 use super::super::{
-    Cell, Coordinates, HexConnectivity, NODE_NUMBERING_OFFSET, NodeMap, Octree, mirror_face,
+    Coordinates, HexConnectivity, NODE_NUMBERING_OFFSET, NodeMap, Octree, face_direction,
 };
-use conspire::math::{TensorRank1, TensorVec, tensor_rank_1};
 
 pub fn apply(
     cells_nodes: &[usize],
     nodes_map: &mut NodeMap,
-    node_index: &mut usize,
     tree: &Octree,
     element_node_connectivity: &mut HexConnectivity,
-    nodal_coordinates: &mut Coordinates,
+    nodal_coordinates: &Coordinates,
 ) {
-    tree.iter().for_each(|cell| {
-        template(
-            0,
-            1,
-            cell,
-            cells_nodes,
-            nodes_map,
-            node_index,
-            tree,
-            element_node_connectivity,
-            nodal_coordinates,
-        );
-        template(
-            0,
-            4,
-            cell,
-            cells_nodes,
-            nodes_map,
-            node_index,
-            tree,
-            element_node_connectivity,
-            nodal_coordinates,
-        );
-        template(
-            1,
-            2,
-            cell,
-            cells_nodes,
-            nodes_map,
-            node_index,
-            tree,
-            element_node_connectivity,
-            nodal_coordinates,
-        );
-        template(
-            1,
-            4,
-            cell,
-            cells_nodes,
-            nodes_map,
-            node_index,
-            tree,
-            element_node_connectivity,
-            nodal_coordinates,
-        );
-        template(
-            2,
-            3,
-            cell,
-            cells_nodes,
-            nodes_map,
-            node_index,
-            tree,
-            element_node_connectivity,
-            nodal_coordinates,
-        );
-        template(
-            2,
-            5,
-            cell,
-            cells_nodes,
-            nodes_map,
-            node_index,
-            tree,
-            element_node_connectivity,
-            nodal_coordinates,
-        );
-        template(
-            3,
-            0,
-            cell,
-            cells_nodes,
-            nodes_map,
-            node_index,
-            tree,
-            element_node_connectivity,
-            nodal_coordinates,
-        );
-        template(
-            4,
-            2,
-            cell,
-            cells_nodes,
-            nodes_map,
-            node_index,
-            tree,
-            element_node_connectivity,
-            nodal_coordinates,
-        );
-        template(
-            4,
-            3,
-            cell,
-            cells_nodes,
-            nodes_map,
-            node_index,
-            tree,
-            element_node_connectivity,
-            nodal_coordinates,
-        );
-        template(
-            5,
-            0,
-            cell,
-            cells_nodes,
-            nodes_map,
-            node_index,
-            tree,
-            element_node_connectivity,
-            nodal_coordinates,
-        );
-        template(
-            5,
-            1,
-            cell,
-            cells_nodes,
-            nodes_map,
-            node_index,
-            tree,
-            element_node_connectivity,
-            nodal_coordinates,
-        );
-        template(
-            3,
-            5,
-            cell,
-            cells_nodes,
-            nodes_map,
-            node_index,
-            tree,
-            element_node_connectivity,
-            nodal_coordinates,
-        );
-    })
-}
-
-#[allow(clippy::too_many_arguments)]
-fn template(
-    face_m: usize,
-    face_n: usize,
-    cell: &Cell,
-    cells_nodes: &[usize],
-    nodes_map: &mut NodeMap,
-    node_index: &mut usize,
-    tree: &Octree,
-    element_node_connectivity: &mut HexConnectivity,
-    nodal_coordinates: &mut Coordinates,
-) {
-    let (
-        subcell_m_a,
-        subcell_m_b,
-        subcell_m_c,
-        subcell_m_d,
-        subcell_face_m_a,
-        subcell_face_m_b,
-        subcell_face_n_a,
-        subcell_face_n_b,
-        subcell_diag_mn_a,
-        subcell_diag_mn_b,
-    ) = match (face_m, face_n) {
-        (0, 1) => (7, 13, 5, 15, 3, 7, 0, 4, 2, 6),
-        (0, 4) => (1, 4, 0, 5, 2, 3, 4, 5, 6, 7),
-        (1, 2) => (7, 13, 5, 15, 2, 6, 1, 5, 0, 4),
-        (1, 4) => (1, 4, 0, 5, 0, 2, 5, 7, 4, 6),
-        (2, 3) => (2, 8, 0, 10, 0, 4, 3, 7, 1, 5),
-        (2, 5) => (11, 14, 10, 15, 4, 5, 2, 3, 0, 1),
-        (3, 0) => (2, 8, 0, 10, 1, 5, 2, 6, 3, 7),
-        (3, 5) => (11, 14, 10, 15, 5, 7, 0, 2, 1, 3),
-        (4, 2) => (11, 14, 10, 15, 6, 7, 0, 1, 4, 5),
-        (4, 3) => (2, 8, 0, 10, 4, 6, 1, 3, 5, 7),
-        (5, 0) => (1, 4, 0, 5, 0, 1, 6, 7, 2, 3),
-        (5, 1) => (7, 13, 5, 15, 1, 3, 4, 6, 0, 2),
-        _ => panic!(),
-    };
-    let directions: [TensorRank1<3, 1>; 2] = [face_m, face_n]
-        .iter()
-        .map(|face| match face {
-            0 => tensor_rank_1([0.0, -1.0, 0.0]),
-            1 => tensor_rank_1([1.0, 0.0, 0.0]),
-            2 => tensor_rank_1([0.0, 1.0, 0.0]),
-            3 => tensor_rank_1([-1.0, 0.0, 0.0]),
-            4 => tensor_rank_1([0.0, 0.0, -1.0]),
-            5 => tensor_rank_1([0.0, 0.0, 1.0]),
-            _ => panic!(),
-        })
-        .collect::<Vec<TensorRank1<3, 1>>>()
-        .try_into()
-        .unwrap();
-    if let Some(cell_face_m) = cell.get_faces()[face_m] {
-        if let Some(cell_face_n) = cell.get_faces()[face_n] {
-            if let Some(cell_diag_mn) = tree[cell_face_m].get_faces()[face_n] {
-                if let Some((subcells_face_m, _)) = tree.cell_contains_leaves(&tree[cell_face_m]) {
-                    if let Some((subcells_face_n, _)) =
-                        tree.cell_contains_leaves(&tree[cell_face_n])
-                    {
-                        if let Some((subcells_diag_mn, _)) =
-                            tree.cell_contains_leaves(&tree[cell_diag_mn])
+    tree.iter()
+        .filter_map(|cell| tree.cell_contains_leaves(cell))
+        .for_each(|(cell_subcells, cell_faces)| {
+            cell_faces
+                .iter()
+                .enumerate()
+                .for_each(|(face_index, face_cell)| {
+                    if let Some(face_cell_index) = face_cell {
+                        if let Some(face_subsubcells) =
+                            tree.cell_subcells_contain_leaves(&tree[*face_cell_index], face_index)
                         {
-                            if let Some(subcells_m) =
-                                tree.cell_subcells_contain_leaves(cell, mirror_face(face_m))
-                            {
-                                if tree
-                                    .cell_subcells_contain_leaves(cell, mirror_face(face_n))
-                                    .is_some()
-                                {
-                                    let lngth = *tree[subcells_m[subcell_m_a]].get_lngth() as f64;
-                                    nodal_coordinates.push(
-                                        &nodal_coordinates[cells_nodes[subcells_m[subcell_m_a]]
-                                            - NODE_NUMBERING_OFFSET]
-                                            + &directions[0] * lngth,
-                                    );
-                                    nodal_coordinates.push(
-                                        &nodal_coordinates[cells_nodes[subcells_m[subcell_m_a]]
-                                            - NODE_NUMBERING_OFFSET]
-                                            + &directions[0] * lngth
-                                            + &directions[1] * lngth,
-                                    );
-                                    nodal_coordinates.push(
-                                        &nodal_coordinates[cells_nodes[subcells_m[subcell_m_a]]
-                                            - NODE_NUMBERING_OFFSET]
-                                            + &directions[1] * lngth,
-                                    );
-                                    nodal_coordinates.push(
-                                        &nodal_coordinates[cells_nodes[subcells_m[subcell_m_b]]
-                                            - NODE_NUMBERING_OFFSET]
-                                            + &directions[0] * lngth,
-                                    );
-                                    nodal_coordinates.push(
-                                        &nodal_coordinates[cells_nodes[subcells_m[subcell_m_b]]
-                                            - NODE_NUMBERING_OFFSET]
-                                            + &directions[0] * lngth
-                                            + &directions[1] * lngth,
-                                    );
-                                    nodal_coordinates.push(
-                                        &nodal_coordinates[cells_nodes[subcells_m[subcell_m_b]]
-                                            - NODE_NUMBERING_OFFSET]
-                                            + &directions[1] * lngth,
-                                    );
-                                    (0..6).for_each(|k| {
-                                        assert!(
-                                            nodes_map
-                                                .insert(
-                                                    (
-                                                        (2.0 * nodal_coordinates[*node_index + k
-                                                            - NODE_NUMBERING_OFFSET][0])
-                                                            as usize,
-                                                        (2.0 * nodal_coordinates[*node_index + k
-                                                            - NODE_NUMBERING_OFFSET][1])
-                                                            as usize,
-                                                        (2.0 * nodal_coordinates[*node_index + k
-                                                            - NODE_NUMBERING_OFFSET][2])
-                                                            as usize,
-                                                    ),
-                                                    *node_index + k,
-                                                )
-                                                .is_none(),
-                                            "duplicate entry"
-                                        )
-                                    });
-                                    element_node_connectivity.push([
-                                        cells_nodes[subcells_m[subcell_m_a]],
-                                        *node_index,
-                                        *node_index + 1,
-                                        *node_index + 2,
-                                        cells_nodes[subcells_m[subcell_m_b]],
-                                        *node_index + 3,
-                                        *node_index + 4,
-                                        *node_index + 5,
-                                    ]);
-                                    element_node_connectivity.push([
-                                        *node_index,
-                                        cells_nodes[subcells_face_m[subcell_face_m_a]],
-                                        cells_nodes[subcells_diag_mn[subcell_diag_mn_a]],
-                                        *node_index + 1,
-                                        *node_index + 3,
-                                        cells_nodes[subcells_face_m[subcell_face_m_b]],
-                                        cells_nodes[subcells_diag_mn[subcell_diag_mn_b]],
-                                        *node_index + 4,
-                                    ]);
-                                    element_node_connectivity.push([
-                                        *node_index + 1,
-                                        cells_nodes[subcells_diag_mn[subcell_diag_mn_a]],
-                                        cells_nodes[subcells_face_n[subcell_face_n_a]],
-                                        *node_index + 2,
-                                        *node_index + 4,
-                                        cells_nodes[subcells_diag_mn[subcell_diag_mn_b]],
-                                        cells_nodes[subcells_face_n[subcell_face_n_b]],
-                                        *node_index + 5,
-                                    ]);
-                                    element_node_connectivity.push([
-                                        cells_nodes[subcells_m[subcell_m_a]],
-                                        *node_index + 2,
-                                        *node_index + 1,
-                                        *node_index,
-                                        cells_nodes[subcells_m[subcell_m_c]],
-                                        cells_nodes[subcells_face_n[subcell_face_n_a]],
-                                        cells_nodes[subcells_diag_mn[subcell_diag_mn_a]],
-                                        cells_nodes[subcells_face_m[subcell_face_m_a]],
-                                    ]);
-                                    element_node_connectivity.push([
-                                        cells_nodes[subcells_m[subcell_m_b]],
-                                        *node_index + 3,
-                                        *node_index + 4,
-                                        *node_index + 5,
-                                        cells_nodes[subcells_m[subcell_m_d]],
-                                        cells_nodes[subcells_face_m[subcell_face_m_b]],
-                                        cells_nodes[subcells_diag_mn[subcell_diag_mn_b]],
-                                        cells_nodes[subcells_face_n[subcell_face_n_b]],
-                                    ]);
-                                    *node_index += 6;
-                                }
-                            }
+                            template_inner(face_index).into_iter().for_each(
+                                |[
+                                    adjacent_face,
+                                    cell_subcell_a,
+                                    cell_subcell_b,
+                                    adjacent_cell_subcell_a,
+                                    adjacent_cell_subcell_b,
+                                    face_subsubcell_a,
+                                    face_subsubcell_b,
+                                    face_subsubcell_c,
+                                    face_subsubcell_d,
+                                    adjacent_face_subsubcell_a,
+                                    adjacent_face_subsubcell_b,
+                                    adjacent_face_subsubcell_c,
+                                    adjacent_face_subsubcell_d,
+                                ]| {
+                                    if let Some(adjacent_cell) = cell_faces[adjacent_face] {
+                                        if let Some((adjacent_cell_subcells, adjacent_cell_faces)) =
+                                            tree.cell_contains_leaves(&tree[adjacent_cell])
+                                        {
+                                            if let Some(adjacent_cell_face_cell) =
+                                                adjacent_cell_faces[face_index]
+                                            {
+                                                if let Some(adjacent_face_subsubcells) = tree
+                                                    .cell_subcells_contain_leaves(
+                                                        &tree[adjacent_cell_face_cell],
+                                                        face_index,
+                                                    )
+                                                {
+                                                    let lngth = *tree
+                                                        [face_subsubcells[face_subsubcell_a]]
+                                                        .get_lngth()
+                                                        as f64;
+                                                    let coordinates_1 = &nodal_coordinates
+                                                        [cells_nodes
+                                                            [face_subsubcells[face_subsubcell_a]]
+                                                            - NODE_NUMBERING_OFFSET]
+                                                        - face_direction(face_index) * lngth;
+                                                    let node_1 = *nodes_map
+                                                        .get(&(
+                                                            (2.0 * coordinates_1[0]) as usize,
+                                                            (2.0 * coordinates_1[1]) as usize,
+                                                            (2.0 * coordinates_1[2]) as usize,
+                                                        ))
+                                                        .expect("nonexistent entry");
+                                                    let coordinates_2 = &nodal_coordinates
+                                                        [cells_nodes
+                                                            [face_subsubcells[face_subsubcell_b]]
+                                                            - NODE_NUMBERING_OFFSET]
+                                                        - face_direction(face_index) * lngth;
+                                                    let node_2 = *nodes_map
+                                                        .get(&(
+                                                            (2.0 * coordinates_2[0]) as usize,
+                                                            (2.0 * coordinates_2[1]) as usize,
+                                                            (2.0 * coordinates_2[2]) as usize,
+                                                        ))
+                                                        .expect("nonexistent entry");
+                                                    let coordinates_3 = &nodal_coordinates
+                                                        [cells_nodes[adjacent_face_subsubcells
+                                                            [adjacent_face_subsubcell_a]]
+                                                            - NODE_NUMBERING_OFFSET]
+                                                        - face_direction(face_index) * lngth;
+                                                    let node_3 = *nodes_map
+                                                        .get(&(
+                                                            (2.0 * coordinates_3[0]) as usize,
+                                                            (2.0 * coordinates_3[1]) as usize,
+                                                            (2.0 * coordinates_3[2]) as usize,
+                                                        ))
+                                                        .expect("nonexistent entry");
+                                                    let coordinates_4 = &nodal_coordinates
+                                                        [cells_nodes[adjacent_face_subsubcells
+                                                            [adjacent_face_subsubcell_b]]
+                                                            - NODE_NUMBERING_OFFSET]
+                                                        - face_direction(face_index) * lngth;
+                                                    let node_4 = *nodes_map
+                                                        .get(&(
+                                                            (2.0 * coordinates_4[0]) as usize,
+                                                            (2.0 * coordinates_4[1]) as usize,
+                                                            (2.0 * coordinates_4[2]) as usize,
+                                                        ))
+                                                        .expect("nonexistent entry");
+                                                    element_node_connectivity.push([
+                                                        cells_nodes[cell_subcells[cell_subcell_a]],
+                                                        cells_nodes[cell_subcells[cell_subcell_b]],
+                                                        node_1,
+                                                        node_2,
+                                                        cells_nodes[adjacent_cell_subcells
+                                                            [adjacent_cell_subcell_a]],
+                                                        cells_nodes[adjacent_cell_subcells
+                                                            [adjacent_cell_subcell_b]],
+                                                        node_3,
+                                                        node_4,
+                                                    ]);
+                                                    element_node_connectivity.push([
+                                                        cells_nodes
+                                                            [face_subsubcells[face_subsubcell_a]],
+                                                        cells_nodes
+                                                            [face_subsubcells[face_subsubcell_b]],
+                                                        node_2,
+                                                        node_1,
+                                                        cells_nodes[adjacent_face_subsubcells
+                                                            [adjacent_face_subsubcell_a]],
+                                                        cells_nodes[adjacent_face_subsubcells
+                                                            [adjacent_face_subsubcell_b]],
+                                                        node_4,
+                                                        node_3,
+                                                    ]);
+                                                    element_node_connectivity.push([
+                                                        cells_nodes
+                                                            [face_subsubcells[face_subsubcell_b]],
+                                                        node_2,
+                                                        node_4,
+                                                        cells_nodes[adjacent_face_subsubcells
+                                                            [adjacent_face_subsubcell_b]],
+                                                        cells_nodes
+                                                            [face_subsubcells[face_subsubcell_d]],
+                                                        cells_nodes[cell_subcells[cell_subcell_a]],
+                                                        cells_nodes[adjacent_cell_subcells
+                                                            [adjacent_cell_subcell_a]],
+                                                        cells_nodes[adjacent_face_subsubcells
+                                                            [adjacent_face_subsubcell_d]],
+                                                    ]);
+                                                    element_node_connectivity.push([
+                                                        cells_nodes
+                                                            [face_subsubcells[face_subsubcell_a]],
+                                                        cells_nodes[adjacent_face_subsubcells
+                                                            [adjacent_face_subsubcell_a]],
+                                                        node_3,
+                                                        node_1,
+                                                        cells_nodes
+                                                            [face_subsubcells[face_subsubcell_c]],
+                                                        cells_nodes[adjacent_face_subsubcells
+                                                            [adjacent_face_subsubcell_c]],
+                                                        cells_nodes[adjacent_cell_subcells
+                                                            [adjacent_cell_subcell_b]],
+                                                        cells_nodes[cell_subcells[cell_subcell_b]],
+                                                    ]);
+                                                }
+                                            }
+                                        }
+                                    }
+                                },
+                            )
                         }
                     }
-                }
-            }
-        }
+                })
+        })
+}
+
+fn template_inner(face_index: usize) -> [[usize; 13]; 2] {
+    match face_index {
+        0 => [
+            [1, 1, 5, 0, 4, 13, 7, 15, 5, 8, 2, 10, 0],
+            [4, 0, 1, 4, 5, 4, 1, 5, 0, 14, 11, 15, 10],
+        ],
+        1 => [
+            [0, 5, 1, 7, 3, 2, 8, 0, 10, 7, 13, 5, 15],
+            [4, 1, 3, 5, 7, 4, 1, 5, 0, 14, 11, 15, 10],
+        ],
+        2 => [
+            [1, 7, 3, 6, 2, 7, 13, 5, 15, 2, 8, 0, 10],
+            [5, 6, 7, 2, 3, 14, 11, 15, 10, 4, 1, 5, 0],
+        ],
+        3 => [
+            [2, 6, 2, 4, 0, 7, 13, 5, 15, 2, 8, 0, 10],
+            [5, 4, 6, 0, 2, 14, 11, 15, 10, 4, 1, 5, 0],
+        ],
+        4 => [
+            [1, 3, 1, 2, 0, 7, 13, 5, 15, 2, 8, 0, 10],
+            [2, 2, 3, 0, 1, 14, 11, 15, 10, 4, 1, 5, 0],
+        ],
+        5 => [
+            [0, 4, 5, 6, 7, 4, 1, 5, 0, 14, 11, 15, 10],
+            [3, 6, 4, 7, 5, 2, 8, 0, 10, 7, 13, 5, 15],
+        ],
+        _ => panic!(),
     }
 }
